pyproject.toml
[project]
name = "memanto"
dynamic = ["version"]
description = "Memory that AI Agents Love!"
authors = [{ name = "Majid Fekri", email = "majid.fekri@edgeaiinnovations.com" }]
requires-python = ">=3.10,<4"
readme = "README.md"
license = "MIT"
keywords = [
    "memanto",
    "moorcheh",
    "memory layer",
    "agentic ai",
    "llm",
    "agent memory",
    "semantic memory",
    "rag",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "moorcheh-sdk>=1.3.5",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "python-multipart>=0.0.7",
]

[project.urls]
Homepage = "https://www.moorcheh.ai"
Documentation = "https://docs.moorcheh.ai"
Repository = "https://github.com/moorcheh-ai/memanto"
"Bug Tracker" = "https://github.com/moorcheh-ai/memanto/issues"

[project.optional-dependencies]
all = [
    "pytest>=8.2.0,<9",
    "pytest-mock>=3.12.0,<4",
    "mypy>=1.10.0,<2",
    "build>=1.2.2.post1,<2",
    "twine>=6.1.0,<7",
    "pre-commit>=4.4.0,<5",
    "ruff>=0.14.6,<0.15",
]

[project.scripts]
memanto = "memanto.cli.main:app"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "memanto/app/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["memanto"]

[dependency-groups]
dev = [
    "pytest>=8.2.0,<9",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0,<4",
    "pytest-timeout>=2.1.0",
    "mypy>=1.10.0,<2",
    "build>=1.2.2.post1,<2",
    "twine>=6.1.0,<7",
    "pre-commit>=4.4.0,<5",
    "ruff>=0.14.6,<0.15",
]

# Optional: Configure MyPy
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# Optional: Configure pytest
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
filterwarnings = [

]

# Optional: Configure Ruff
[tool.ruff]
target-version = "py310"
line-length = 88
extend-exclude = ["legacy_archive"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "B904",  # raise exception without from
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]
